[M] Fix gesture response fault on iOS devices
When building a player UI based on VlcPlayer, the screen won't be able to respond to tap gestures when calling or hiding self implemented widget before VlcPlayer widget in a Stack widget. Add this param should solve the problem. I believe the maintainers must have forgotten to add this since they already added it in AndroidView.